requirements.txt
atlassian-python-api==3.31.0
Jinja2==3.1.2
mkdocs==1.4.1
mkdocs-awesome-pages-plugin==2.8.0
mkdocs-exclude==1.0.2
mkdocs-material==8.5.7
mkdocs-localsearch==0.9.2
mkdocs-material-extensions==1.1
PyYAML==6.0
tqdm==4.64.1
docopt==0.6.2
cpe==1.2.1
